Put debug layout functions into gtc_backend.common
